Comments from Dmitry HW4

+++ src/main/java/homeworks/hw4/SelenideSliderPage.java
@@ -9,20 +9,24 @@ import static org.testng.Assert.assertTrue;
 
 public class SelenideSliderPage {
 
-    // TODO It is completely prohibited to create locators in this way !
+    // TODO It is completely prohibited to create locators in this way ! x2
 
+    // TODO Once again, this locators should be improved
     @FindBy(css = "[class='ui-slider-handle ui-state-default ui-corner-all']")
     private ElementsCollection sliders;
 
     @FindBy(css = "[class='panel-body-list logs']")
     private SelenideElement logs;
 
+    // TODO Once again, this locators should be improved
     @FindBy(css = "[class='uui-slider blue range ui-slider ui-slider-horizontal ui-widget ui-widget-content ui-corner-all']")
     private SelenideElement slider;
 
+    // TODO Once again, this locators should be improved
     @FindBy(css = "[class='uui-slider blue range ui-slider ui-slider-horizontal ui-widget ui-widget-content ui-corner-all'] > a:nth-child(3)")
     private SelenideElement rightSlider;
 
+    // TODO Once again, this locators should be improved
     @FindBy(css = "[class='uui-slider blue range ui-slider ui-slider-horizontal ui-widget ui-widget-content ui-corner-all'] > a:nth-child(1)")
     private SelenideElement leftSlider;
 
@@ -31,12 +35,12 @@ public class SelenideSliderPage {
     // TODO Take a look on Math.round
     // TODO Peace of hint: you can organise your algorithm around current values of slider, it will be easier.
 
-
     public void moveSlidersBoundryPosition(int left, int right) {
         assertCorrectValues(left, right);
         actions().clickAndHold(leftSlider).moveByOffset(-1000, 0).release().build().perform();
         actions().clickAndHold(rightSlider).moveByOffset(-1000, 0).release().build().perform();
-        int step = Math.round(slider.getSize().width / 100);
+        // TODO Pay attention, 100 != 100.0
+        int step = (int) Math.round(slider.getSize().width / 100.0);
         int actualRightValue = Integer.valueOf(leftSlider.getText());
         int actualLeftValue = Integer.valueOf(leftSlider.getText());
         if (right == 100) {
